Extract email handler policies in separate files
