+++ components/email-service/src/Altinn.Notifications.Email.Integrations/Wolverine/Handlers/CheckEmailSendStatusHandler.cs
@@ -1,20 +1,11 @@
-using System.Diagnostics.CodeAnalysis;
-
 using Altinn.Notifications.Email.Core;
 using Altinn.Notifications.Email.Core.Configuration;
 using Altinn.Notifications.Email.Core.Dependencies;
 using Altinn.Notifications.Email.Core.Models;
 using Altinn.Notifications.Email.Core.Status;
-using Altinn.Notifications.Email.Integrations.Configuration;
-
-using Azure.Messaging.ServiceBus;
-
-using Microsoft.Extensions.Options;
 
 using Wolverine;
 using Wolverine.Attributes;
-using Wolverine.ErrorHandling;
-using Wolverine.Runtime.Handlers;
 
 namespace Altinn.Notifications.Email.Integrations.Wolverine.Handlers;
 
@@ -25,32 +16,6 @@ namespace Altinn.Notifications.Email.Integrations.Wolverine.Handlers;
 public static class CheckEmailSendStatusHandler
 {
     private const int _statusPollDelayMs = 8000;
-
-    /// <summary>
-    /// Configures error-handling and retry policies for the polling-loop handler chain.
-    /// Wolverine resolves <paramref name="options"/> from the IoC container at chain-compilation time.
-    /// </summary>
-    /// <remarks>
-    /// <see cref="ArgumentException"/> is intentionally excluded from the retry policy.
-    /// Messages that fail guard-clause validation are malformed and must not be retried — they
-    /// will be moved to the dead-letter queue by Wolverine's default failure handling.
-    /// </remarks>
-    /// <param name="chain">The Wolverine handler chain to apply the retry policies to.</param>
-    /// <param name="options">The Wolverine settings resolved from the IoC container.</param>
-    [ExcludeFromCodeCoverage]
-    public static void Configure(HandlerChain chain, IOptions<WolverineSettings> options)
-    {
-        var policy = options.Value.EmailStatusCheckQueuePolicy;
-
-        chain
-            .OnException<TimeoutException>()
-            .Or<ServiceBusException>()
-            .Or<TaskCanceledException>()
-            .Or<InvalidOperationException>()
-            .RetryWithCooldown(policy.GetCooldownDelays())
-            .Then.ScheduleRetry(policy.GetScheduleDelays())
-            .Then.MoveToErrorQueue();
-    }
 
     /// <summary>
     /// Polls ACS for delivery status. If the result is terminal, publishes
